feat: Add gilded rose kata and related files
This commit adds the gilded rose kata and its related files. The gilded rose kata is a famous exercise that involves refactoring code to make it easier to understand and maintain. The code has been copied from the original repository.

Resources related to the gilded rose kata can be found in the respective README file.